# Cache `mmMaxTime` in a map instead of `memSeries`