--- a/tsdb/head_wal.go
+++ b/tsdb/head_wal.go
@@ -479,13 +479,11 @@ func (h *Head) resetSeriesWithMMappedChunks(mSeries *memSeries, mmc, oooMmc []*m
 		}
 		*mSeries.ooo = memSeriesOOOFields{oooMmappedChunks: oooMmc}
 	}
-	// Cache the last mmapped chunk time, so we can skip calling append() for samples it will reject.
-	if len(mmc) == 0 {
-		mSeries.mmMaxTime = math.MinInt64
-	} else {
-		mSeries.mmMaxTime = mmc[len(mmc)-1].maxTime
-		h.updateMinMaxTime(mmc[0].minTime, mSeries.mmMaxTime)
+
+	if len(mmc) > 0 {
+		h.updateMinMaxTime(mmc[0].minTime, mmc[len(mmc)-1].maxTime)
 	}
+
 	if len(oooMmc) != 0 {
 		// Mint and maxt can be in any chunk, they are not sorted.
 		mint, maxt := int64(math.MaxInt64), int64(math.MinInt64)
@@ -569,12 +567,18 @@ func (wp *walSubsetProcessor) processWALSamples(h *Head, mmappedChunks, oooMmapp
 		samplesPerChunk: h.opts.SamplesPerChunk,
 	}
 
+	// Cache the last mmapped chunk time, so we can skip calling append() for samples it will reject.
+	mmMaxTimes := map[chunks.HeadSeriesRef]int64{}
+
 	for in := range wp.input {
 		if in.existingSeries != nil {
 			mmc := mmappedChunks[in.walSeriesRef]
 			oooMmc := oooMmappedChunks[in.walSeriesRef]
 			if h.resetSeriesWithMMappedChunks(in.existingSeries, mmc, oooMmc, in.walSeriesRef) {
 				mmapOverlappingChunks++
+			}
+			if len(mmc) > 0 {
+				mmMaxTimes[in.existingSeries.ref] = mmc[len(mmc)-1].maxTime
 			}
 			continue
 		}
@@ -585,7 +589,7 @@ func (wp *walSubsetProcessor) processWALSamples(h *Head, mmappedChunks, oooMmapp
 				unknownRefs++
 				continue
 			}
-			if s.T <= ms.mmMaxTime {
+			if mmMaxTime, ok := mmMaxTimes[s.Ref]; ok && s.T <= mmMaxTime {
 				continue
 			}
 			if _, chunkCreated := ms.append(s.T, s.V, 0, appendChunkOpts); chunkCreated {
@@ -614,7 +618,7 @@ func (wp *walSubsetProcessor) processWALSamples(h *Head, mmappedChunks, oooMmapp
 				unknownHistogramRefs++
 				continue
 			}
-			if s.t <= ms.mmMaxTime {
+			if mmMaxTime, ok := mmMaxTimes[s.ref]; ok && s.t <= mmMaxTime {
 				continue
 			}
 			var chunkCreated bool