@colega 2024-07-16
I'd like to benchmark this with a real WAL just like Bryan did in https://github.com/prometheus/prometheus/pull/9160