@colega 2024-07-29
main:
![image](https://github.com/user-attachments/assets/39b7f351-b335-4b88-9cdf-c23deee71f4b)

cache:
![image](https://github.com/user-attachments/assets/a54649f9-1a68-44fa-afc4-d7b6c3a909db)

Note the extra map lookup.