@colega 2024-07-29
Benchmark data size:
```
$ du -h
0	./wbl
376M	./wal/checkpoint.00078035
5.4G	./wal
6.3G	./chunks_head
12G	.
```

I'm on an M3 Pro Mac, and these are the results:
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
               │    main    │               cache               │
               │   sec/op   │   sec/op     vs base              │
LoadRealWLs-12   99.82 ± 6%   108.33 ± 4%  +8.52% (p=0.009 n=6)

               │     main     │               cache                │
               │     B/op     │     B/op      vs base              │
LoadRealWLs-12   34.99Gi ± 0%   35.84Gi ± 0%  +2.43% (p=0.002 n=6)

               │    main     │               cache               │
               │  allocs/op  │  allocs/op   vs base              │
LoadRealWLs-12   382.8M ± 0%   383.2M ± 0%  +0.12% (p=0.002 n=6)
```